Add addresses with validator/tests

+++ app/Http/Controllers/AddressController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\Address;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 
 class AddressController extends Controller
 {
@@ -16,9 +18,9 @@ class AddressController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
-        return Address::all();
+        return $request->user()->addresses;
     }
 
     /**
@@ -29,7 +31,19 @@ class AddressController extends Controller
      */
     public function store(Request $request)
     {
-        $user = Address::create([
+        $validator = Validator::make($request->all(), [
+            'country' => 'required|string',
+            'city' => 'required|string',
+            'postal_code' => 'required|string|max:5|min:5',
+            'street' => 'required|string',
+            'type' => 'in:idk1,idk2'
+        ]);
+
+        if ($validator->fails()) {
+            return $validator->getMessageBag();
+        }
+
+        $adress = Address::create([
             'country' => $request->country,
             'city' => $request->city,
             'postal_code' => $request->postal_code,
@@ -38,18 +52,40 @@ class AddressController extends Controller
             'user_id' => $request->user()->id
         ]);
 
-        return $user;
+        return $adress;
     }
 
     /**
      * Display the specified resource.
      *
+     * @param  \Illuminate\Http\Request  $request
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
+    public function show(Request $request, $id)
     {
-        return Address::find($id);
+        $adress = Address::find($id);
+        $user = $request->user();
+
+        if (is_null($adress)) {
+            $data = [
+                'message' => "L'adresse n'existe pas !"
+            ];
+            $code_response = 404;
+        } elseif ($user->id !== $adress->user_id) {
+            $data = [
+                'message' => "L'adresse ne vous appartient pas !"
+            ];
+            $code_response = 403;
+        } else {
+            $data = [
+                'data' => $adress,
+                'message' => null
+            ];
+            $code_response = 200;
+        }
+
+        return response()->json($data, $code_response);
     }
 
     /**
@@ -61,18 +97,71 @@ class AddressController extends Controller
      */
     public function update(Request $request, $id)
     {
-        Address::find($id)->update($request->all());
-        return Address::find($id);
+        $adress = Address::find($id);
+        $user = $request->user();
+
+        $validator = Validator::make($request->all(), [
+            'country' => 'required|string',
+            'city' => 'required|string',
+            'postal_code' => 'required|string|max:5|min:5',
+            'street' => 'required|string',
+            'type' => 'in:idk1,idk2',
+            'user_id' => 'nullable'
+        ]);
+        if ($validator->fails()) {
+            return $validator->getMessageBag();
+        }
+
+        if (is_null($adress)) {
+            $data = [
+                'message' => "L'adresse n'existe pas !"
+            ];
+            $code_response = 404;
+        } elseif ($user->id !== $adress->user_id) {
+            $data = [
+                'message' => "L'adresse ne vous appartient pas !"
+            ];
+            $code_response = 403;
+        } else {
+            $adress->update($request->all());
+            $data = [
+                'data' => $adress,
+                'message' => "La modification à bien été effectuée"
+            ];
+            $code_response = 200;
+        }
+
+        return response()->json($data, $code_response);
     }
 
     /**
      * Remove the specified resource from storage.
-     *
+     * 
+     * @param  \Illuminate\Http\Request  $request
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
-        Address::find($id)->delete();
+        $adress = Address::find($id);
+        $user = $request->user();
+        if (is_null($adress)) {
+            $data = [
+                'message' => "L'adresse n'existe pas !"
+            ];
+            $code_response = 404;
+        } elseif ($user->id !== $adress->user_id) {
+            $data = [
+                'message' => "L'adresse ne vous appartient pas !"
+            ];
+            $code_response = 403;
+        } else {
+            Address::find($id)->delete();
+            $data = [
+                'message' => "La supression à bien été effectuée"
+            ];
+            $code_response = 200;
+        }
+        return response()->json($data, $code_response);
     }
 }
